requirements.txt
ollama
chromadb
rich
pypdf
python-docx
beautifulsoup4
markdown
pandas
requests
reportlab
